--- mdd/codec.go
@@ -1,6 +1,0 @@
-package mdd
-
-type Codec interface {
-	Decode([]byte) (*Containers, error)
-	Encode(*Containers) ([]byte, error)
-}
