+++ mdd/codec/type.go
@@ -1,0 +1,9 @@
+package codec
+
+type Type int
+
+const (
+	Unknown Type = iota
+	CMDC
+	BMDC
+)
